critical core onsip bundle that interfaces with the jitsi applet
this is the most important addition to the jitsi project. it is the custom bundle
that wraps all the capabilities of the web phone and interfaces with the applet.
it includes volume control, a feature added later in the development process.
note: some necessary fixed had to be made when upgrading from revision 8587.