removing applet failover logic that's no longer applicable
java crashes when putting the computer into a hibernate state.
the previous codebase kept credentials on the filesystem so
that when the applet restarted, these credentials were retrievable
and could be used to re-register the account that was loaded
prior to the crash.  this is no longer the case, so we've
modified the recovery flow, and this code is no longer necessary.