configuration update: allows end user to control ring volume or use default
a slight enhancement to the control of the ringer volume. affords the api the
ability to set the ringer volume in sync with playback volume or keep the
ringer volume at some default while allowing the playback volume to still
be controlled by the end user.

tagged: release-1.2 in web-jitsi-lib repo